bug fix HTC-402

+++ src/main/java/kz/dilau/htcdatamanager/config/CommissionRange.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 @Getter
 @Setter
 public class CommissionRange {
-    private int from;
-    private int to;
+    private long from;
+    private long to;
     private BigDecimal amount;
 }
